+++ src/MCEq/__init__.py
@@ -1,4 +1,4 @@
-import mceq_config as config
+import src.mceq_config as config
 import MCEq.version
 
 __version__ = MCEq.version.__version__
