+++ src/MCEq/geometry/nrlmsise00/DOCUMENTATION
